+++ dantro/groups/__init__.py
@@ -1,5 +1,5 @@
 """The groups sub-package implements BaseDataGroup specializations"""
 
-from .ordered import OrderedDataGroup
+from .ordered import OrderedDataGroup, IndexedDataGroup
 from .pspgrp import ParamSpaceStateGroup, ParamSpaceGroup
 from .network import NetworkGroup