+++ dantro/mixins/__init__.py
@@ -6,5 +6,6 @@ from .base import AttrsMixin, PathMixin, CheckDataMixin
 from .base import CollectionMixin, ItemAccessMixin, MappingAccessMixin
 
 from .general import ForwardAttrsToDataMixin
+from .indexing import IntegerItemAccessMixin, PaddedIntegerItemAccessMixin
 from .numeric import UnaryOperationsMixin, NumbersMixin, ComparisonMixin
 from .proxy_support import ProxySupportMixin, Hdf5ProxySupportMixin